requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset_normalizer==2.0.4
click==8.1.3
cloudpickle==1.6.0
dask[distributed]==2022.6.0
decorator==5.0.9
distributed==2022.6.0
dm-tree==0.1.6
flatbuffers==2.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
gin==0.1.006
gin-config==0.4.0
google-auth-oauthlib==0.4.5
google-auth==1.35.0
google-pasta==0.2.0
grpcio==1.39.0
gym==0.19.0
HeapDict==1.0.1
h5py==3.6.0
idna==3.2
Jinja2==3.1.2
keras==2.7.0
keras-preprocessing==1.1.2
libclang==12.0.0
locket==1.0.0
markdown==3.3.4
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.22.1
oauthlib==3.1.1
opt-einsum==3.3.0
partd==1.2.0
pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1_modules==0.2.8
pyglet==1.5.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
setuptools==57.4.0
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-probability==0.15.0
termcolor==1.1.0
tf-agents==0.11.0
toolz==0.11.2
tornado==6.1
typing-extensions==4.0.1
urllib3==1.26.6
werkzeug==2.0.1
wheel==0.37.0
wrapt==1.13.3
zict==2.2.0
zipp==3.7.0
